MVP-6295-patch-3: Fix build static files issue
