+++ packages/notifi-react-example-v2/src/app/link/[slug]/page.tsx
@@ -1,59 +1,14 @@
-'use client';
-
-import {
-  NotifiSmartLink,
-  NotifiSmartLinkContextProvider,
-  useNotifiSmartLinkContext,
-} from '@notifi-network/notifi-react';
+import { NotifiSmartLinkExample } from '@/components/NotifiSmartLinkExample';
 import '@notifi-network/notifi-react/dist/index.css';
-import { useWallets } from '@notifi-network/notifi-wallet-provider';
-import { usePathname } from 'next/navigation';
 import React from 'react';
 
+export async function generateStaticParams() {
+  // Replace this with your supported SmartLink IDs
+  return [
+    { slug: '66881b08-189b-4276-9ca0-9cd6fe099b4a' },
+    { slug: '0f624c59e09a4d2183a79bfc95177c67' },
+  ];
+}
 export default function Page() {
-  const { wallets, selectedWallet } = useWallets();
-  const { authParams } = useNotifiSmartLinkContext();
-  const currentPath = usePathname();
-
-  const smartLinkId = currentPath.split('/').pop();
-  console.log('smartLinkId', smartLinkId);
-  if (!smartLinkId) {
-    return <div>Smart Link ID is required</div>;
-  }
-  return (
-    <div>
-      <div>
-        <h3>Smart Link Example: metamask</h3>
-        <NotifiSmartLinkContextProvider
-          env={'Development'}
-          authParams={{
-            walletPublicKey:
-              selectedWallet === 'metamask'
-                ? (wallets[selectedWallet].walletKeys?.hex ?? '')
-                : '',
-            walletBlockchain: 'ARBITRUM',
-          }}
-        >
-          <NotifiSmartLink
-            smartLinkId={smartLinkId}
-            preAction={{
-              isRequired: selectedWallet !== 'metamask',
-              onClick: async () => {
-                if (!selectedWallet) {
-                  wallets['metamask'].connect();
-                }
-              },
-              label: 'Connect EVM Wallet',
-            }}
-            actionHandler={async (args) => {
-              console.log('Action triggered (react-example-v2)', {
-                args,
-                authParams,
-              });
-            }}
-          />
-        </NotifiSmartLinkContextProvider>
-      </div>
-    </div>
-  );
+  return <NotifiSmartLinkExample />;
 }
